# CheckCheck/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CheckCheck REST API",
    "version": "0.1.dev47+ga91abbb.d20260618"
  },
  "paths": {
...
    "/api/auth/basic/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth Basic Register",
        "operationId": "auth_basic_register_api_auth_basic_register_post",
        "parameters": [
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          },
          {
            "name": "password_repeat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password Repeat"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "title": "Email"
            }
          },
          {
            "name": "display_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Display Name"
            }
          },
          {
            "name": "user_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 128,
              "pattern": "^[a-zA-Z0-9.-]+$",
              "title": "User Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/basic/login/session": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth Basic Login Session Based",
        "description": "Cookie based session login.",
        "operationId": "auth_basic_login_session_based_api_auth_basic_login_session_post",
        "parameters": [
          {
            "name": "target_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicLoginBody"
              }
            }
          }
        },
        "responses": {
          "307": {
            "description": "Set session cookie and redirect to `target_path` page or start page."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/basic/login/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth Basic Login Token Based",
        "operationId": "auth_basic_login_token_based_api_auth_basic_login_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_auth_basic_login_token_based_api_auth_basic_login_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIToken"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/oidc/login/{provider_slug}/session": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth Oidc Login Session Based",
        "operationId": "auth_oidc_login_session_based_api_auth_oidc_login__provider_slug__session_get",
        "parameters": [
          {
            "name": "provider_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Slug"
            }
          },
          {
            "name": "target_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Path"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Redirect to external OpenIDConnect provider authentikation endpoint. . After coming back the callback endpoint will set a session cookie and redirect to the webapp."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/oidc/login/{provider_slug}/token": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth Oidc Login Token Based",
        "operationId": "auth_oidc_login_token_based_api_auth_oidc_login__provider_slug__token_get",
        "parameters": [
          {
            "name": "provider_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Slug"
            }
          },
          {
            "name": "target_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Path"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Redirect to external OpenIDConnect provider authentikation endpoint. After coming back the callback endpoint will return an API access token."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/oidc/callback/{provider_slug}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Auth Oidc Callback",
        "description": "Callback endpoint triggered after redirection from the OpenID Connect (OIDC) provider.<br><br>\nDepending on the login method used:<br>\n- **Session login** (`/auth/oidc/login/{provider_slug}/session`) returns a **307 Redirect** to the `target_path`.<br>\n- **Token login** (`/auth/oidc/login/{provider_slug}/token`) returns a **200 OK** with a JSON access token.<br><br>\nSupports both browser-based and API-based flows.",
        "operationId": "auth_oidc_callback_api_auth_oidc_callback__provider_slug__get",
        "parameters": [
          {
            "name": "provider_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return an API access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIToken"
                }
              }
            }
          },
          "307": {
            "description": "Set a session cookie and redirect to the `target_path` given to the login endpoint.",
            "content": {
              "application/json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIToken": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "APIToken"
      },
...
      "BasicLoginBody": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "BasicLoginBody"
      },
      "Body_auth_basic_login_token_based_api_auth_basic_login_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_auth_basic_login_token_based_api_auth_basic_login_token_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}